requirements.txt
# python 3.7.7
selenium==4.5.0
webdriver-manager==3.8.3
Pillow==9.2.0